server/requirements.txt
Django==3.0.8

gunicorn==20.0.4 # http сервер, который свзявает nginx с django приложением

ipdb==0.13.3 # отладка программ
ipython==7.12.0 # приятная работа в shell
django-debug-toolbar==2.2 # панель для отладки запросов в бд

Pillow==7.2.0 # для работы с изображениями
djangorestframework==3.11.0 # API
django-url-filter==0.3.15 # улучшенные фильтры для DRF

# для селери
celery==4.4.1
flower==0.9.3
redis==3.4.1
requests==2.24.0

channels-redis==3.0.1
channels==2.4.0
daphne==2.5.0

# тестирование
pytest-django==3.9.0
pytest-cov==2.10.0
pytest-flake8==1.0.6

django-fsm
gunicorn==20.0.4